+++ .github/workflows/python-app.yml
@@ -37,6 +37,6 @@ jobs:
         pip install wheel setuptools
         # Build
         python setup.py sdist bdist_wheel
-    #- name: Test with pytest
-    #  run: |
-    #    pytest --cov=mfasweep test/
+    - name: Test with pytest
+      run: |
+        export PYTHONPATH='.';export MFASWEEP_PASSWORD='password'; pytest --cov mfasweep --verbose --cov-config .coveragerc
