+++ tests/test_imports.py
@@ -1,0 +1,5 @@
+
+def test_imports():
+    import requests
+    import colorama
+    import mfasweep
